backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.2
scikit-learn==1.5.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pydantic==2.8.2
pydantic-settings==2.4.0
transformers==4.43.3
torch>=2.2.0
python-multipart==0.0.9

# --- New: For LangChain & Vector DB Integration ---
langchain==0.2.14
langchain-community==0.2.12
pinecone-client==5.0.1
huggingface-hub>=0.23.0

# --- Optional: For OpenAI GenAI Support (via LangChain) ---
openai>=1.42.0
